+++ src/cli/src/commands/lint.ts
@@ -23,7 +23,7 @@ import {
     calculateElapsedNanoseconds,
     formatElapsedNanosecondsAsMilliseconds,
     readMonotonicNanoseconds
-} from "../shared/timing/elapsed-time.js";
+} from "../shared/timing/verbose-timing.js";
 import { formatPathForDisplay } from "../workflow/display-path.js";
 import {
     discoverProjectRoot,
