+++ src/cli/src/shared/timing/index.ts
@@ -1,2 +1,1 @@
-export * from "./elapsed-time.js";
 export * from "./verbose-timing.js";
